+++ tests/tests_algorithms/test_catboost.py
@@ -126,6 +126,7 @@ class CatBoostAlgorithmTest(unittest.TestCase):
         y_predicted = cat.predict(self.X)
         loss = metric(self.y, y_predicted)
 
+        backup_name = None
         try: 
             with tempfile.NamedTemporaryFile() as tmp:
                 #Save and reload using temporary file
@@ -147,8 +148,10 @@ class CatBoostAlgorithmTest(unittest.TestCase):
 
         #Ensure we clean up after ourselves.
         finally:
-            if os.path.exists(backup_name): os.remove(backup_name)
-            if os.path.exists(tmp.name): os.remove(tmp.name)
+            if backup_name is not None:
+                if os.path.exists(backup_name): os.remove(backup_name)
+            if tmp.name is not None:
+                if os.path.exists(tmp.name): os.remove(tmp.name)
 
         #Compare predictions between both models
         y_predicted = cat2.predict(self.X)
